+++ bin/model_generator.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:path/path.dart';
+import 'package:meta/meta.dart';
 
 import 'src/config/pubspec_config.dart';
 import 'src/config/yml_generator_config.dart';
@@ -26,25 +27,26 @@ Future<void> main(List<String> args) async {
         'This program requires a config file. `model_generator/config.yaml`');
   }
   final modelGeneratorContent = configFile.readAsStringSync();
-  final modelGeneratorConfig = YmlGeneratorConfig(modelGeneratorContent);
+  final modelGeneratorConfig =
+      YmlGeneratorConfig(pubspecConfig, modelGeneratorContent);
 
   writeToFiles(pubspecConfig, modelGeneratorConfig);
-  await generateJsonGeneratedModels();
+  await generateJsonGeneratedModels(useFvm: pubspecConfig.useFvm);
   print('Done!!!');
 }
 
 void writeToFiles(
     PubspecConfig pubspecConfig, YmlGeneratorConfig modelGeneratorConfig) {
-  final modelDirectory = Directory(join('lib', 'model'));
-  if (!modelDirectory.existsSync()) {
-    modelDirectory.createSync(recursive: true);
-  }
   modelGeneratorConfig.models.forEach((model) {
+    final modelDirectory = Directory(join('lib', model.baseDirectory));
+    if (!modelDirectory.existsSync()) {
+      modelDirectory.createSync(recursive: true);
+    }
     String content;
     if (model is ObjectModel) {
-      content = ObjectModelWriter(pubspecConfig.projectName, model).write();
+      content = ObjectModelWriter(pubspecConfig, model).write();
     } else if (model is EnumModel) {
-      content = EnumModelWriter(pubspecConfig.projectName, model).write();
+      content = EnumModelWriter(model).write();
     }
     if (model is! CustomModel && content == null) {
       throw Exception(
@@ -52,9 +54,10 @@ void writeToFiles(
     }
     File file;
     if (model.path == null) {
-      file = File(join('lib', 'model', '${model.fileName}.dart'));
+      file = File(join('lib', model.baseDirectory, '${model.fileName}.dart'));
     } else {
-      file = File(join('lib', 'model', model.path, '${model.fileName}.dart'));
+      file = File(join(
+          'lib', model.baseDirectory, model.path, '${model.fileName}.dart'));
     }
     if (!file.existsSync()) {
       file.createSync(recursive: true);
@@ -62,29 +65,32 @@ void writeToFiles(
 
     if (model is! CustomModel) {
       file.writeAsStringSync(content);
-
-      File generatedFile;
-      if (model.path == null) {
-        generatedFile = File(join('lib', 'model', '${model.fileName}.g.dart'));
-      } else {
-        generatedFile =
-            File(join('lib', 'model', model.path, '${model.fileName}.g.dart'));
-      }
-      generatedFile.writeAsStringSync("part of '${model.fileName}.dart';");
     }
   });
 }
 
-/// run `flutter packages pub run build_runner build --delete-conflicting-outputs`
-Future<void> generateJsonGeneratedModels() async {
-  final result = Process.runSync('flutter', [
-    'packages',
-    'pub',
-    'run',
-    'build_runner',
-    'build',
-    '--delete-conflicting-outputs',
-  ]);
+Future<void> generateJsonGeneratedModels({@required bool useFvm}) async {
+  ProcessResult result;
+  if (useFvm) {
+    result = Process.runSync('fvm', [
+      'flutter',
+      'packages',
+      'pub',
+      'run',
+      'build_runner',
+      'build',
+      '--delete-conflicting-outputs',
+    ]);
+  } else {
+    result = Process.runSync('flutter', [
+      'packages',
+      'pub',
+      'run',
+      'build_runner',
+      'build',
+      '--delete-conflicting-outputs',
+    ]);
+  }
   if (result.exitCode == 0) {
     print('Succesfully generated the jsonSerializable generated files');
     print('');