+++ bin/src/config/yml_generator_config.dart
@@ -1,8 +1,5 @@
 import 'package:yaml/yaml.dart';
 
-import '../model/model/custom_from_to_json_model.dart';
-import '../model/model/custom_model.dart';
-import '../model/model/enum_model.dart';
 import '../model/field.dart';
 import '../model/item_type/array_type.dart';
 import '../model/item_type/boolean_type.dart';
@@ -13,26 +10,32 @@ import '../model/item_type/integer_type.dart';
 import '../model/item_type/item_type.dart';
 import '../model/item_type/object_type.dart';
 import '../model/item_type/string_type.dart';
+import '../model/model/custom_from_to_json_model.dart';
+import '../model/model/custom_model.dart';
+import '../model/model/enum_model.dart';
 import '../model/model/model.dart';
 import '../model/object_model.dart';
 import '../util/type_checker.dart';
+import 'pubspec_config.dart';
 
 class YmlGeneratorConfig {
   static final _models = <Model>[];
 
   List<Model> get models => _models;
 
-  YmlGeneratorConfig(String configContent) {
+  YmlGeneratorConfig(PubspecConfig pubspecConfig, String configContent) {
     loadYaml(configContent).forEach((key, value) {
+      final String baseDirectory =
+          value['base_directory'] ?? pubspecConfig.baseDirectory;
       final String path = value['path'];
       final YamlMap properties = value['properties'];
       final String type = value['type'];
       if (type == 'custom') {
-        models.add(CustomModel(key, path));
+        models.add(CustomModel(key, path, baseDirectory));
         return;
       }
       if (type == 'custom_from_to_json') {
-        models.add(CustomFromToJsonModel(key, path));
+        models.add(CustomFromToJsonModel(key, path, baseDirectory));
         return;
       }
       if (properties == null) {
@@ -47,7 +50,7 @@ class YmlGeneratorConfig {
           fields.add(EnumField(propertyKey,
               propertyValue == null ? '' : propertyValue['value']));
         });
-        models.add(EnumModel(key, path, fields));
+        models.add(EnumModel(key, path, baseDirectory, fields));
       } else {
         final fields = <Field>[];
         properties.forEach((propertyKey, propertyValue) {
@@ -56,12 +59,12 @@ class YmlGeneratorConfig {
           }
           fields.add(getField(propertyKey, propertyValue));
         });
-        models.add(ObjectModel(key, path, fields));
+        models.add(ObjectModel(key, path, baseDirectory, fields));
       }
     });
 
     checkIfTypesAvailable();
-    addPathsToFields();
+    addPathsToFields(pubspecConfig);
   }
 
   Field getField(String name, YamlMap property) {
@@ -70,6 +73,10 @@ class YmlGeneratorConfig {
           property.containsKey('required') && property['required'] == true;
       final ignored =
           property.containsKey('ignore') && property['ignore'] == true;
+      final includeIfNull = property.containsKey('includeIfNull') &&
+          property['includeIfNull'] == false;
+      final nonFinal = ignored ||
+          property.containsKey('non_final') && property['non_final'] == true;
       final jsonKey = property['jsonKey'] ?? property['jsonkey'];
       final type = property['type'];
       ItemType itemType;
@@ -112,25 +119,35 @@ class YmlGeneratorConfig {
         itemType = ObjectType(ref);
       }
       return Field(
-          name: name,
-          type: itemType,
-          required: required,
-          ignore: ignored,
-          jsonKey: jsonKey);
+        name: name,
+        type: itemType,
+        required: required,
+        ignore: ignored,
+        jsonKey: jsonKey,
+        nonFinal: nonFinal,
+        includeIfNull: includeIfNull,
+      );
     } catch (e) {
       print('Something went wrong with $name:\n\n${e.toString()}');
       rethrow;
     }
   }
 
-  void addPathsToFields() {
+  void addPathsToFields(PubspecConfig pubspecConfig) {
     models.forEach((model) {
       if (model is ObjectModel) {
         model.fields.forEach((field) {
           final foundModels =
               models.where((model) => model.name == field.type.name).toList();
           if (foundModels.isNotEmpty) {
-            field.path = foundModels[0].path;
+            final foundModel = foundModels[0];
+            final baseDirectory =
+                foundModel.baseDirectory ?? pubspecConfig.baseDirectory;
+            if (foundModel.path == null) {
+              field.path = '$baseDirectory';
+            } else {
+              field.path = '$baseDirectory/${foundModel.path}';
+            }
           }
         });
       } else if (model is EnumModel) {}