+++ example/lib/model/user/testing.g.dart
@@ -12,6 +12,7 @@ Testing _$TestingFromJson(Map<String, dynamic> json) {
     beneficiary: json['beneficiary'] as String,
     structuredMessage: json['structuredMessage'] as String,
     beneficiaryIBAN: json['beneficiaryIBAN'] as String,
+    dynamicField: json['dynamicField'],
     duration: json['duration'] == null
         ? null
         : Duration.fromJson(json['duration'] as Map<String, dynamic>),
@@ -24,6 +25,7 @@ Map<String, dynamic> _$TestingToJson(Testing instance) => <String, dynamic>{
       'beneficiary': instance.beneficiary,
       'structuredMessage': instance.structuredMessage,
       'beneficiaryIBAN': instance.beneficiaryIBAN,
+      'dynamicField': instance.dynamicField,
       'duration': instance.duration,
       'duration_from_json_test':
           handleDurationFromToJsonToJson(instance.durationFromJsonTest),